requirements.txt
azure-functions
requests
pandas
pyodbc